Add files via upload
As an infrastructure in the transport system, Alternative Fuel Stations are an important part of the popularisation of new energy vehicles. The purpose of this analysis is to find out the factors affecting the popularity of new energy vehicles by studying the layout of Alternative Fuel Stations in the United States. It is also hoped that the existing Alternative Fuel Stations can be used to predict the future development trend of Alternative Fuel Stations, so as to give some suggestions to encourage the popularity of new energy vehicles and reduce the problem of climate warming. Spatial analysis was used to analyse the site layout. In predicting the development trend, time series prediction was used.